+++ src/compas_viewer/scene/__init__.py
@@ -47,11 +47,11 @@ from .cylinderobject import CylinderObject
 from .ellipseobject import EllipseObject
 from .coneobject import ConeObject
 from .capsuleobject import CapsuleObject
+from .polyhedronobject import PolyhedronObject
 from .geometryobject import GeometryObject
 from .groupobject import Group
 from .groupobject import GroupObject
 from .collectionobject import Collection
-from .polyhedronobject import PolyhedronObject
 from .collectionobject import CollectionObject
 
 
@@ -87,9 +87,9 @@ def register_scene_objects():
     register(Ellipse, EllipseObject, context="Viewer")
     register(Cone, ConeObject, context="Viewer")
     register(Capsule, CapsuleObject, context="Viewer")
+    register(Polyhedron, PolyhedronObject, context="Viewer")
     register(list, GroupObject, context="Viewer")
     register(Collection, CollectionObject, context="Viewer")
-    register(Polyhedron, PolyhedronObject, context="Viewer")
 
     try:
         from compas_occ.brep import OCCBrep